+++ packages/basic/src/basic/useBasic.ts
@@ -1,9 +1,9 @@
-import { useAtomValue } from '@einfach/react'
+import { useAtomValue, Store } from '@einfach/react'
 import type {} from '@einfach/react'
 import type { Core } from './state'
 import { basicAtom } from './state'
 
-export function useBasic(): Core {
-  const res = useAtomValue(basicAtom)
+export function useBasic(store?: Store): Core {
+  const res = useAtomValue(basicAtom, { store })
   return res
 }