+++ packages/table/src/core/useDataInit.ts
@@ -1,13 +1,14 @@
-import { useEffect } from 'react'
+import { useEffect, useMemo } from 'react'
 import { useAtomValue, useSetAtom } from '@einfach/react'
 import type { DataItem, UseDataProps } from '../types'
 import { useData } from './useData'
-import { useBasic } from '@grid-table/basic'
+import { RowId, useBasic } from '@grid-table/basic'
 import { useExpand } from '../tree'
 import './useDataInit.css'
 import { dataInitAtom } from '../state'
 import { columnInitAtom } from '../stateColumn'
 import { headerDataInitAtom } from '../stateHeader'
+import { getColumnId } from '../utils/getColumnId'
 
 export function useDataInit<ItemInfo extends DataItem>(props: UseDataProps<ItemInfo>) {
   const { dataSource, columns } = props
@@ -58,7 +59,31 @@ export function useDataInit<ItemInfo extends DataItem>(props: UseDataProps<ItemI
   }, [initHeaderData, props.headerDataSource, props.rowHeight])
   useExpand()
 
+  /** 固定列功能 */
+  const { stickyList } = useMemo(() => {
+    const leftFixedColList: RowId[] = []
+    const rightFixedColList: RowId[] = []
+
+    columns.forEach((column, index) => {
+      const columnId = getColumnId(column)
+      if (column.fixed === 'left') {
+        leftFixedColList.push(columnId)
+      }
+      if (column.fixed === 'right') {
+        rightFixedColList.push(columnId)
+      }
+    })
+    return {
+      columnCount: columns.length,
+      stickyList: {
+        topIdList: leftFixedColList,
+        bottomIdList: rightFixedColList,
+      },
+    }
+  }, [columns])
+
   return {
     loading,
+    stickyList,
   }
 }